+++ reflex-dom/java/org/reflexfrp/reflexdom/MainWidget.java
@@ -41,13 +41,8 @@ public class MainWidget {
     a.setContentView(wv);
     final WebSettings ws = wv.getSettings();
     ws.setJavaScriptEnabled(true);
-    ws.setAllowFileAccessFromFileURLs(true);
-    ws.setAllowUniversalAccessFromFileURLs(true);
     ws.setDomStorageEnabled(true);
     wv.setWebContentsDebuggingEnabled(true);
-    // allow file access on newer android
-    ws.setAllowContentAccess(true);
-    ws.setAllowFileAccess(true);
     // allow video to play without user interaction
     wv.getSettings().setMediaPlaybackRequiresUserGesture(false);
     final AtomicBoolean jsaddleLoaded = new AtomicBoolean(false);
@@ -67,7 +62,8 @@ public class MainWidget {
         @Override
         public WebResourceResponse shouldInterceptRequest (WebView view, WebResourceRequest request) {
             Uri uri = request.getUrl();
-            if(!uri.getScheme().equals("file"))
+	    Log.i("reflex", uri.toString());
+            if(!uri.getScheme().equals("file") && !uri.getEncodedAuthority().equals("appassets.androidplatform.net"))
                 return null;
 
             String path = uri.getPath();